divravb

+++ ByteCodeRun.cpp
@@ -641,15 +641,15 @@ inline bool ByteCodeRun::executeInstruction(ByteCodeRunContext* context, ByteCod
 
     case ByteCodeOp::MulRAVB:
     {
-        registersRC[ip->a.u32].s32 *= ip->b.s32;
+        registersRC[ip->a.u32].s32 *= ip->b.u32;
         break;
     }
     case ByteCodeOp::DivRAVB:
     {
-        if (ip->b.s32 == 0)
+        if (ip->b.u32 == 0)
             context->error("division by zero");
         else
-            registersRC[ip->a.u32].s32 /= ip->b.s32;
+            registersRC[ip->a.u32].s64 /= ip->b.u32;
         break;
     }
